handlebars creation can be overridden by overriding HandlebarsFactory, close #429
